Post create error fixed by refactoring the code
